src/main/java/com/sergiomartinrubio/tddbyexample/controller/TransactionController.java
package com.sergiomartinrubio.tddbyexample.controller;

import com.sergiomartinrubio.tddbyexample.model.Transaction;
import com.sergiomartinrubio.tddbyexample.service.TransactionService;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/transaction")
public class TransactionController {

    private final TransactionService transactionService;

    public TransactionController(TransactionService transactionService) {
        this.transactionService = transactionService;
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public void save(@RequestBody Transaction transaction) {
        transactionService.save(transaction);
    }
}
